+++ tests/elfs/ticks.py
@@ -4,7 +4,7 @@ from .common import case
 CASES = [
     case(
         name="ticks",
-        command="runelf apps/tests/ticks",
+        command="runelf apps/tests/ticks alpha beta",
         must_contain=[
             "ticks program",
             "ticks = ",